requirements.txt
aiogram
python-dotenv==0.21.0
redis
aioredis==2.0.1
pymongo==4.2.0
mongoengine==0.24.2
yarl
babel==2.10.3
